move getImageData and getFileUrl to utils file
